+++ SparseGrids/tsgCudaMacros.hpp
@@ -54,7 +54,7 @@ namespace TasGrid{
 namespace TasCUDA{
     // general GPU/CPU I/O, new/delete vectors, send/recv data
     template <typename T>
-    inline T* cudaNew(int num_entries, std::ostream *os = 0){
+    inline T* cudaNew(size_t num_entries, std::ostream *os = 0){
         T* x = 0;
         cudaError_t cudaStat = cudaMalloc(((void**) &x), num_entries * sizeof(T));
         #ifdef _TASMANIAN_DEBUG_
@@ -64,7 +64,7 @@ namespace TasCUDA{
     }
 
     template <typename T>
-    inline T* cudaSend(int num_entries, const T *cpu_array, std::ostream *os = 0){
+    inline T* cudaSend(size_t num_entries, const T *cpu_array, std::ostream *os = 0){
         T *x = cudaNew<T>(num_entries, os);
         cudaError_t cudaStat = cudaMemcpy(x, cpu_array, num_entries * sizeof(T), cudaMemcpyHostToDevice);
         #ifdef _TASMANIAN_DEBUG_
@@ -74,7 +74,7 @@ namespace TasCUDA{
     }
 
     template <typename T>
-    inline void cudaSend(int num_entries, const T *cpu_array, T *gpu_array, std::ostream *os = 0){
+    inline void cudaSend(size_t num_entries, const T *cpu_array, T *gpu_array, std::ostream *os = 0){
         cudaError_t cudaStat = cudaMemcpy(gpu_array, cpu_array, num_entries * sizeof(T), cudaMemcpyHostToDevice);
         #ifdef _TASMANIAN_DEBUG_
         AccelerationMeta::cudaCheckError((void*) &cudaStat, "cudaSend(type, type)", os);
@@ -82,7 +82,7 @@ namespace TasCUDA{
     }
 
     template <typename T>
-    inline void cudaRecv(int num_entries, const T *gpu_array, T *cpu_array, std::ostream *os = 0){
+    inline void cudaRecv(size_t num_entries, const T *gpu_array, T *cpu_array, std::ostream *os = 0){
         cudaError_t cudaStat = cudaMemcpy(cpu_array, gpu_array, num_entries * sizeof(T), cudaMemcpyDeviceToHost);
         #ifdef _TASMANIAN_DEBUG_
         AccelerationMeta::cudaCheckError((void*) &cudaStat, "cudaRecv(type, type)", os);
